refactor: services 경로 → shared 경로 변경
