+++ src/components/Icon/IconPlus.tsx
@@ -1,6 +1,6 @@
 import { ComponentProps } from 'react';
 
-import { tw } from '@/services/utils/tailwindMerge';
+import { tw } from '@/shared/utils/tailwindMerge';
 
 const IconPlus = ({ className, ...props }: ComponentProps<'svg'>) => (
   <svg
