+++ src/components/Input/TextField/TextField.tsx
@@ -1,6 +1,6 @@
 import { ComponentPropsWithRef, PropsWithChildren } from 'react';
 
-import { tw } from '@/services/utils/tailwindMerge';
+import { tw } from '@/shared/utils/tailwindMerge';
 import IconPassword from '@/components/Icon/IconPassword';
 import Tag from '@/components/Tag/Tag';
